+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 2.2.0
 
 - Bump: sentry-android to v4.0.0 #1309
 - build(ios): Bump sentry-cocoa to 6.1.4 #1308